+++ tests/api-resources/workflow-recipient-runs.test.ts
@@ -1,0 +1,53 @@
+// File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
+
+import Knock from '@knocklabs/node';
+
+const client = new Knock({
+  apiKey: 'My API Key',
+  baseURL: process.env['TEST_API_BASE_URL'] ?? 'http://127.0.0.1:4010',
+});
+
+describe('resource workflowRecipientRuns', () => {
+  test('list', async () => {
+    const responsePromise = client.workflowRecipientRuns.list();
+    const rawResponse = await responsePromise.asResponse();
+    expect(rawResponse).toBeInstanceOf(Response);
+    const response = await responsePromise;
+    expect(response).not.toBeInstanceOf(Response);
+    const dataAndResponse = await responsePromise.withResponse();
+    expect(dataAndResponse.data).toBe(response);
+    expect(dataAndResponse.response).toBe(rawResponse);
+  });
+
+  test('list: request options and params are passed correctly', async () => {
+    // ensure the request options are being passed correctly by passing an invalid HTTP method in order to cause an error
+    await expect(
+      client.workflowRecipientRuns.list(
+        {
+          after: 'after',
+          before: 'before',
+          ending_at: '2019-12-27T18:11:19.117Z',
+          has_errors: true,
+          page_size: 0,
+          recipient: 'user_123',
+          starting_at: '2019-12-27T18:11:19.117Z',
+          status: ['queued'],
+          tenant: 'tenant',
+          workflow: 'workflow',
+        },
+        { path: '/_stainless_unknown_path' },
+      ),
+    ).rejects.toThrow(Knock.NotFoundError);
+  });
+
+  test('get', async () => {
+    const responsePromise = client.workflowRecipientRuns.get('id');
+    const rawResponse = await responsePromise.asResponse();
+    expect(rawResponse).toBeInstanceOf(Response);
+    const response = await responsePromise;
+    expect(response).not.toBeInstanceOf(Response);
+    const dataAndResponse = await responsePromise.withResponse();
+    expect(dataAndResponse.data).toBe(response);
+    expect(dataAndResponse.response).toBe(rawResponse);
+  });
+});